parse all dwarven structures

+++ src/OpenSage.Game/Logic/Object/Draw/W3dStreakDraw.cs
@@ -11,7 +11,7 @@ namespace OpenSage.Logic.Object
         private static readonly IniParseTable<W3dStreakDrawModuleData> FieldParseTable = new IniParseTable<W3dStreakDrawModuleData>
         {
             { "Length", (parser, x) => x.Length =  parser.ParseInteger() },
-            { "Width", (parser, x) => x.Width = parser.ParseInteger() },
+            { "Width", (parser, x) => x.Width = parser.ParseFloat() },
             { "NumSegments", (parser, x) => x.NumSegments = parser.ParseInteger() },
             { "Color", (parser, x) => x.Color = parser.ParseColorRgba() },
             { "Texture", (parser, x) => x.Texture = parser.ParseAssetReference() },
@@ -20,7 +20,7 @@ namespace OpenSage.Logic.Object
         };
 
         public int Length { get; private set; }
-        public int Width { get; private set; }
+        public float Width { get; private set; }
         public int NumSegments { get; private set; }
         public ColorRgba Color { get; private set; }
         public string Texture { get; private set; }
